requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
python-dotenv>=1.0
pydantic>=2.7
requests>=2.32